Update README.md
##The idea is to:
Detect malware using:
*Generic Algorithm based intrusion detection: GA-IDS (Spoofed IPs)
*File extension
*Threat intelligence cloud
*Encrypted Unique Bin identifiers co-joined with GA-IDS
it will check many things in the network including: 1. File extension 2. Threat intelligence cloud community around the world 3. hashes of and wrong format a. GA-IDS and our encryption is the only the DUT gets information. The ransomware will not be encrypted inside the DUT and won't be using our encryption either so the bin unique is going to be different. 
##Prevention:
If we can identify them we can analyze them in real time to prevent them by immobilizing them completely using DatashiledAI encryption.